AppBar's Cascading Value needs set Name
The "Fixed" cascading value in AppBar is not named which causes issues down the line. A child component of AppBar can take the RightToLeft as a CascadingParameter and will take the Fixed cascading value in AppBar as the value since it is not named and of the same type. An example of this would be a MudBadge inside of the AppBar. It's RightToLeft parameter would take the value of the Fixed cascading value.